# specs/001-build-a-ridesharing/contracts/api-spec.yaml
openapi: 3.0.0
info:
  title: AlbaniaRides API
  description: Ridesharing platform API for connecting drivers and passengers in Albania
  version: 1.0.0
  contact:
    name: AlbaniaRides Support
    email: support@albaniarides.com
servers:
  - url: https://api.albaniarides.com/v1
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 2
          maxLength: 100
        phone:
          type: string
          pattern: '^\+355[0-9]{8,9}$'
        city:
          type: string
        photo_url:
          type: string
          format: uri
        bio:
          type: string
          maxLength: 500
        is_driver:
          type: boolean
        car_model:
          type: string
        car_color:
          type: string
        driving_years:
          type: integer
          minimum: 0
        rating:
          type: number
          minimum: 1
          maximum: 5
        total_rides:
          type: integer
        member_since:
          type: string
          format: date-time
    Ride:
      type: object
      required:
        - origin_city
        - destination_city
        - departure_time
        - seats_total
        - price_per_seat
        - pickup_point
      properties:
        id:
          type: string
          format: uuid
        driver:
          $ref: '#/components/schemas/UserSummary'
        origin_city:
          type: string
          enum: [Tirana, Durrës, Vlorë, Shkodër, Elbasan, Fier, Korçë, Berat, Lushnjë, Kavajë, Pogradec, Gjirokastër, Sarandë, Laç, Kukës]
        destination_city:
          type: string
          enum: [Tirana, Durrës, Vlorë, Shkodër, Elbasan, Fier, Korçë, Berat, Lushnjë, Kavajë, Pogradec, Gjirokastër, Sarandë, Laç, Kukës]
        pickup_point:
          type: string
        departure_time:
          type: string
          format: date-time
        seats_total:
          type: integer
          minimum: 1
          maximum: 4
        seats_available:
          type: integer
          minimum: 0
          maximum: 4
        price_per_seat:
          type: number
          minimum: 0
        stops:
          type: array
          items:
            type: string
        luggage_space:
          type: boolean
        smoking_allowed:
          type: boolean
        is_recurring:
          type: boolean
        recurrence_pattern:
          type: string
          enum: [daily, weekly]
        status:
          type: string
          enum: [active, completed, cancelled]
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        photo_url:
          type: string
          format: uri
        rating:
          type: number
        total_rides:
          type: integer
...
paths:
  /auth/register:
    post:
      summary: Register new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - name
                - city
              properties:
                phone:
                  type: string
                  pattern: '^\+355[0-9]{8,9}$'
                name:
                  type: string
                city:
                  type: string
      responses:
        '201':
          description: Registration successful, SMS sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  verification_required:
                    type: boolean
        '400':
          description: Invalid input
        '409':
          description: Phone number already registered
  /auth/verify:
    post:
      summary: Verify phone number with OTP
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - otp
              properties:
                phone:
                  type: string
                otp:
                  type: string
      responses:
        '200':
          description: Verification successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid OTP
        '404':
          description: Phone number not found
  /auth/login:
    post:
      summary: Login with phone number
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
              properties:
                phone:
                  type: string
                  pattern: '^\+355[0-9]{8,9}$'
      responses:
        '200':
          description: SMS sent
        '404':
          description: User not found
  /rides:
    get:
      summary: Search for rides
      tags: [Rides]
      parameters:
        - name: origin
          in: query
          required: true
          schema:
            type: string
        - name: destination
          in: query
          required: true
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: time_period
          in: query
          schema:
            type: string
            enum: [morning, afternoon, evening]
        - name: sort
          in: query
          schema:
            type: string
            enum: [departure, price, rating]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 20
      responses:
        '200':
          description: List of rides
          content:
            application/json:
              schema:
                type: object
                properties:
                  rides:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ride'
                  total:
                    type: integer
                  page:
                    type: integer
                  pages:
                    type: integer

    post:
      summary: Create a new ride
      tags: [Rides]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ride'
      responses:
        '201':
          description: Ride created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ride'
        '401':
          description: Unauthorized
        '400':
          description: Invalid input
...